requirements.txt
appdirs==1.4.4
asgiref==3.3.4
attr==0.3.1
attrs==19.1.0
autopep8==1.5.4
azure-core==1.14.0
azure-storage-blob==12.8.1
blis==0.7.4
boto==2.49.0
boto3==1.16.28
botocore==1.19.28
boxing==0.1.4
cachetools==4.2.2
catalogue==2.0.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
cymem==2.0.5
Cython==0.29.23
Django==3.1.4
django-annoying==0.10.6
django-cors-headers==3.6.0
django-debug-toolbar==3.2
django-extensions==3.1.0
django-filter==2.4.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-redis-cache==3.0.0
django-rest-swagger==2.2.0
django-rq==2.3.2
django-user-agents==0.4.0
djangorestframework==3.12.4
drf-dynamic-fields==0.3.0
drf-generators==0.3.0
drf-yasg==1.20.0
Flask==1.1.2
google-api-core==1.29.0
google-auth==1.30.1
google-cloud-core==1.6.0
google-cloud-storage==1.28.1
google-resumable-media==0.5.1
googleapis-common-protos==1.53.0
htmlmin==0.1.12
idna==2.10
inflection==0.5.1
isodate==0.6.0
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
jsonschema==3.2.0
label-studio==1.0.2.post0
label-studio-converter==0.0.28
lockfile==0.12.2
lxml==4.6.3
MarkupSafe==1.1.1
mock==4.0.3
more-itertools==8.8.0
moto==1.3.16.dev122
msrest==0.6.21
murmurhash==1.0.5
nltk==3.5
numpy==1.20.3
oauthlib==3.1.1
openapi-codec==1.3.2
ordered-set==4.0.2
packaging==20.9
pandas==1.2.4
pathy==0.5.2
Pillow==7.2.0
preshed==3.0.5
protobuf==3.17.2
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.7.3
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.4.1
redis==3.5.3
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
responses==0.13.3
river==0.7.1
rq==1.7.0
rsa==4.7.2
ruamel.yaml==0.17.7
ruamel.yaml.clib==0.2.2
rules==2.2
s3transfer==0.3.7
scikit-learn==0.24.1
scipy==1.6.3
simplejson==3.17.2
six==1.16.0
smart-open==3.0.0
spacy==3.0.6
spacy-legacy==3.0.5
sqlparse==0.4.1
srsly==2.4.1
thinc==8.0.3
threadpoolctl==2.1.0
torch==1.8.1
torchvision==0.9.1
tqdm==4.61.0
typer==0.3.2
typing-extensions==3.10.0.0
ua-parser==0.10.0
ujson==4.0.2
uritemplate==3.0.1
urllib3==1.26.5
user-agents==2.2.0
wasabi==0.8.2
Werkzeug==2.0.1
wincertstore==0.2
xmljson==0.2.0
xmltodict==0.12.0
zipp==3.4.1
